requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.12=h4ab18f5_0
aom=3.9.1=hac33072_0
blinker=1.8.2=pyhd8ed1ab_0
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.33.1=heb4867d_0
ca-certificates=2024.8.30=hbcca054_0
cairo=1.18.0=hebfffa5_3
certifi=2024.8.30=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
contourpy=1.3.0=py312h68727a3_2
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hd590300_0
dbus=1.13.6=h5008d03_3
double-conversion=3.3.0=h59595ed_0
expat=2.6.3=h5888daf_0
ffmpeg=7.0.2=gpl_h3fac18e_105
flask=3.0.3=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_2
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.54.1=py312h66e93f0_0
freeglut=3.2.2=ha6d2627_3
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h36c2ea0_0
gmp=6.3.0=hac33072_2
graphite2=1.3.13=h59595ed_1003
harfbuzz=9.0.0=hda332d3_1
hdf5=1.14.3=nompi_hdf9ad27_105
icu=75.1=he02047a_0
imath=3.1.12=h7955e40_0
importlib-metadata=8.5.0=pyha770c72_0
itsdangerous=2.2.0=pyhd8ed1ab_0
jasper=4.2.4=h536e39c_0
jinja2=3.1.4=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py312h68727a3_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_1
lerc=4.0.0=h27087fc_0
libabseil=20240116.2=cxx17_he02047a_1
libaec=1.1.3=h59595ed_0
libasprintf=0.22.5=he8f35ee_3
libass=0.17.3=h1dc1e6a_0
libblas=3.9.0=24_linux64_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=24_linux64_openblas
libclang-cpp19.1=19.1.0=default_hb5137d0_0
libclang13=19.1.0=default_h9c6a7e4_0
libcups=2.3.3=h4637d8d_4
libcurl=8.10.1=hbbe4b11_0
libdeflate=1.21=h4bc722e_0
libdrm=2.4.123=hb9d3cd8_0
libedit=3.1.20191231=he28a2e2_2
libegl=1.7.0=ha4b6fd6_0
libev=4.33=hd590300_2
libexpat=2.6.3=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgfortran=14.1.0=h69a702a_1
libgfortran-ng=14.1.0=h69a702a_1
libgfortran5=14.1.0=hc5f4f2c_1
libgl=1.7.0=ha4b6fd6_0
libglib=2.82.1=h2ff4ddf_0
libglu=9.0.0=ha6d2627_1004
libglvnd=1.7.0=ha4b6fd6_0
libglx=1.7.0=ha4b6fd6_0
libgomp=14.1.0=h77fa898_1
libhwloc=2.11.1=default_hecaa2ac_1000
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=24_linux64_openblas
liblapacke=3.9.0=24_linux64_openblas
libllvm19=19.1.0=ha7bfdaf_0
libnghttp2=1.58.0=h47da74e_1
libnsl=2.0.1=hd590300_0
libopenblas=0.3.27=pthreads_hac2b453_1
libopencv=4.10.0=qt6_py312h01efb12_604
libopengl=1.7.0=ha4b6fd6_0
libopenvino=2024.4.0=hac27bb2_0
libopenvino-auto-batch-plugin=2024.4.0=h4d9b6c2_0
libopenvino-auto-plugin=2024.4.0=h4d9b6c2_0
libopenvino-hetero-plugin=2024.4.0=h3f63f65_0
libopenvino-intel-cpu-plugin=2024.4.0=hac27bb2_0
libopenvino-intel-gpu-plugin=2024.4.0=hac27bb2_0
libopenvino-intel-npu-plugin=2024.4.0=hac27bb2_0
libopenvino-ir-frontend=2024.4.0=h3f63f65_0
libopenvino-onnx-frontend=2024.4.0=h56242b0_0
libopenvino-paddle-frontend=2024.4.0=h56242b0_0
libopenvino-pytorch-frontend=2024.4.0=h5888daf_0
libopenvino-tensorflow-frontend=2024.4.0=h358ae18_0
libopenvino-tensorflow-lite-frontend=2024.4.0=h5888daf_0
libopus=1.3.1=h7f98852_1
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libpq=16.4=h2d7952a_1
libprotobuf=4.25.3=hd5b35b9_1
libsqlite=3.46.1=hadc24fc_0
libssh2=1.11.0=h0841786_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libtiff=4.7.0=h6565414_0
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h8a09558_1
libvpx=1.14.1=hac33072_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.17.0=h8a09558_0
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.12.7=he7c6b58_4
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=h4ab18f5_1
markupsafe=2.1.5=py312h66e93f0_1
matplotlib=3.9.2=py312h7900ff3_1
matplotlib-base=3.9.2=py312hd3ec401_1
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=9.0.1=h266115a_1
mysql-libs=9.0.1=he0572af_1
ncurses=6.5=he02047a_1
numpy=2.1.1=py312h58c1407_0
ocl-icd=2.3.2=hd590300_1
opencv=4.10.0=qt6_py312hac6a15e_604
openexr=3.2.2=h04e0de5_2
openh264=2.4.1=h59595ed_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=py312hf9745cd_1
pcre2=10.44=hba22ea6_2
pillow=10.4.0=py312h56024de_1
pip=24.2=pyh8b19718_1
pixman=0.43.2=h59595ed_0
pthread-stubs=0.4=hb9d3cd8_1002
pugixml=1.14=h59595ed_0
py-opencv=4.10.0=qt6_py312h071dcc1_604
pyparsing=3.1.4=pyhd8ed1ab_0
pyside6=6.7.2=py312h91f0f75_4
python=3.12.0=hab00c5b_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2024.2=pyhd8ed1ab_0
python_abi=3.12=5_cp312
pytz=2024.1=pyhd8ed1ab_0
qhull=2020.2=h434a139_5
qt6-main=6.7.2=hadfd74e_5
readline=8.2=h8228510_1
setuptools=75.1.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.2.1=ha2e4443_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
svt-av1=2.2.1=h5888daf_0
tbb=2021.13.0=h84d6215_0
tk=8.6.13=noxft_h4845f30_101
tornado=6.4.1=py312h66e93f0_1
tzdata=2024a=h8827d51_1
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.37=hd8ed1ab_0
werkzeug=3.0.4=pyhd8ed1ab_0
wheel=0.44.0=pyhd8ed1ab_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.42=h4ab18f5_0
xorg-fixesproto=5.0=hb9d3cd8_1003
xorg-inputproto=2.3.2=hb9d3cd8_1003
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.10=h4f16b4b_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxdmcp=1.1.3=hb9d3cd8_1
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h4bc722e_1
xorg-libxrender=0.9.11=hd590300_0
xorg-libxtst=1.2.5=h4bc722e_0
xorg-libxxf86vm=1.1.5=hb9d3cd8_2
xorg-recordproto=1.14.2=hb9d3cd8_1003
xorg-renderproto=0.11.1=hb9d3cd8_1003
xorg-xextproto=7.3.0=hb9d3cd8_1004
xorg-xorgproto=2024.1=hb9d3cd8_1
xz=5.2.6=h166bdaf_0
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.3.1=h4ab18f5_1
zstd=1.5.6=ha6fb4c9_0
